requirements.txt
streamlit>=1.31.0
langchain>=0.2.0
langgraph>=0.1.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-google-genai>=1.0.0
langchain-chroma>=0.1.0
langchain-huggingface>=0.0.0
langchain-text-splitters>=0.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
tavily-python>=0.3.0
pypdf>=4.0.0
sentence-transformers>=2.3.0
chromadb>=0.4.20
ragas==0.2.1
datasets==3.2.0
sentence-transformers>=2.0.0
accelerate>=1.1.0
